Desktop/깃하이라이트/requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.12.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==2.3.8
mysqlclient==2.1.1
pymysql==1.1.0
gunicorn==21.2.0
